crewai-backend/requirements.txt
crewai>=0.80.0
crewai-tools>=0.13.0
langchain-mistralai>=0.1.12
python-dotenv>=1.0.1
requests>=2.31.0
requests-cache>=1.2.0
beautifulsoup4>=4.12.3
pydantic>=2.7.0
